+++ src/main/java/org/synyx/urlaubsverwaltung/security/config/SecurityLdapConfigurationProperties.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Component;
 @ConfigurationProperties("uv.security.ldap")
 public class SecurityLdapConfigurationProperties {
 
+    private boolean enabled;
+
     private String url;
     private String base;
     private String managerDn;
@@ -15,6 +17,14 @@ public class SecurityLdapConfigurationProperties {
     private String userSearchFilter;
 
     private SecurityLdapSync sync = new SecurityLdapSync();
+
+    public boolean getEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
 
     public String getUrl() {
         return url;